Update Kernel and remove offsets calculation. Move emitters cpp.
