+++ src/common/snippets/include/snippets/op/kernel.hpp
@@ -28,7 +28,6 @@ public:
     std::shared_ptr<Node> clone_with_new_inputs(const OutputVector& inputs) const override {
         return std::make_shared<Kernel>(region);
     }
-    const void *compile_params = nullptr;
 };
 
 } // namespace op
